Renamed files to reflect the class name
