+++ applications/GeoMechanicsApplication/custom_elements/updated_lagrangian_U_Pw_axisymmetric_FIC_element.cpp
@@ -12,7 +12,7 @@
 
 // Application includes
 #include "custom_elements/updated_lagrangian_U_Pw_axisymmetric_FIC_element.hpp"
-#include "axisymmetric_stress_state_policy.h"
+#include "axisymmetric_stress_state.h"
 
 namespace Kratos
 {
