+++ applications/GeoMechanicsApplication/custom_elements/updated_lagrangian_U_Pw_diff_order_axisymmetric_element.cpp
@@ -12,7 +12,7 @@
 
 // Application includes
 #include "custom_elements/updated_lagrangian_U_Pw_diff_order_axisymmetric_element.hpp"
-#include "axisymmetric_stress_state_policy.h"
+#include "axisymmetric_stress_state.h"
 
 namespace Kratos
 {
